fix(refactor): Decouple business logic from VisualizationStepViews (#1183)
